style(firmware): :art: format (#18)

+++ firmware/motor_test.cydsn/main.c
@@ -13,8 +13,8 @@
 #include "project.h"
 #include "vars.h"
 
-#include <stdio.h>
 #include <stdbool.h>
+#include <stdio.h>
 
 void usbPutString(char *s);
 void usbPutChar(char c);
@@ -51,8 +51,8 @@ volatile float battery_reading;
 volatile bool data_ready = 0;
 CY_ISR(SenseBattery)
 {
-    battery_reading = (ADC_SAR_1_CountsTo_Volts(ADC_SAR_1_GetResult16()) * 3.518);
-    data_ready = 1;
+	battery_reading = (ADC_SAR_1_CountsTo_Volts(ADC_SAR_1_GetResult16()) * 3.518);
+	data_ready = 1;
 }
 
 int main(void)
@@ -66,8 +66,8 @@ int main(void)
 	QuadDec_M1_Start();
 	QuadDec_M2_Start();
 	isr_speed_StartEx(CalculateSpeed);
-    Timer_ADC_Start();
-    ADC_SAR_1_Start();
+	Timer_ADC_Start();
+	ADC_SAR_1_Start();
 	isr_bat_StartEx(SenseBattery);
 
 	PWM_M1_WriteCompare(0);
@@ -77,12 +77,13 @@ int main(void)
 	USBUART_Start(0, USBUART_5V_OPERATION);
 #endif
 
-    while(data_ready == 0){
-        ;
-    }
-    sprintf(bat_entry, "Battery: %dmV\n", (uint)(battery_reading*1000));
-    usbPutString(bat_entry);
-    ADC_SAR_1_Stop();
+	while (data_ready == 0)
+	{
+		;
+	}
+	sprintf(bat_entry, "Battery: %dmV\n", (uint16_t)(battery_reading * 1000));
+	usbPutString(bat_entry);
+	ADC_SAR_1_Stop();
 	for (;;)
 	{
 		/* Place your application code here. */
@@ -110,7 +111,6 @@ int main(void)
 			}
 			PWM_M1_WriteCompare(PWM_MAX * (dutyCycle / (float)100));
 			PWM_M2_WriteCompare(PWM_MAX * (dutyCycle / (float)100));
-            
 		}
 	}
 }
